Moved sidebar children to sidebar component
